MAT-6223: Update TestCaseTable to TanStack
