+++ package.json
@@ -99,7 +99,7 @@
     "@mui/x-date-pickers": "^6.6.0",
     "@tailwindcss/forms": "^0.5.7",
     "@tanstack/match-sorter-utils": "^8.8.4",
-    "@tanstack/react-table": "^8.9.3",
+    "@tanstack/react-table": "^8.10.3",
     "ace-builds": "~1.4.13",
     "allotment": "^1.19.0",
     "axios": "^1.6.1",
@@ -110,6 +110,7 @@
     "file-saver": "^2.0.5",
     "formik": "^2.2.9",
     "fqm-execution": "^1.3.3",
+    "html-react-parser": "^1.4.12",
     "jszip": "^3.10.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
@@ -119,7 +120,6 @@
     "react-dom": "^17.0.2",
     "react-dropzone": "^14.2.3",
     "react-router": "^6.2.1",
-    "html-react-parser": "^1.4.12",
     "react-router-dom": "^6.2.1",
     "single-spa": "^5.9.3",
     "single-spa-react": "^4.4.0",
