+++ src/components/testCaseLanding/qdm/TestCaseList.tsx
@@ -25,7 +25,7 @@ import {
   TestCasesPassingDetailsProps,
   TestCaseListProps,
 } from "../common/interfaces";
-import TestCaseTable from "../common/TestCaseTable";
+import TestCaseTable from "../common/TestCaseTable/TestCaseTable";
 import UseTestCases from "../common/Hooks/UseTestCases";
 import UseToast from "../common/Hooks/UseToast";
 import { useQdmExecutionContext } from "../../routes/qdm/QdmExecutionContext";
@@ -141,7 +141,7 @@ const TestCaseList = (props: TestCaseListProps) => {
     ) {
       setSelectedPopCriteria(measure.groups[0]);
       const newPath = `/measures/${measureId}/edit/test-cases/list-page/${measure.groups[0].id}`;
-      navigate(newPath, { replace: true }); // update route
+      navigate(newPath);
       if (
         measure?.errors?.length > 0 &&
         (measure.errors.includes(
